feat(cluster): extract BEPEngine/Provider/Model into internal/engine (Phase 2 S1)
Moves BEPEngine, BEPProvider, AgentSyncModel, PeerConnection, and the
receiver ring buffer from root package main into internal/engine so the
daemon (cmd/cogos) can eventually drive cross-node sync without importing
root. All BEP types now live canonically in pkg/substrate/bep; both
packages consume them through type aliases. The SetBus(*busSessionManager)
coupling is inverted to a func(bep.SyncEvent) callback (SetEventCallback).

Root package main is reduced to thin shims (type aliases + forwarding
constructors). Both binaries build; daemon stays GOOS=windows-clean.
Part of #330 (S1) — no daemon startup wiring (S2).
